(#112) refactor: 액티비티 페이지 경로 변경
